modified I2C_bus

+++ lib/I2C_bus/I2C_bus.h
@@ -45,7 +45,7 @@ uint8_t I2C_bus<T_I2C_bus>::read_register(uint8_t address, uint8_t reg)
 	bus->beginTransmission(address);
 	bus->write(reg);
 	bus->endTransmission();
-	bus->requestFrom(address, 1);
+	bus->requestFrom((uint8_t)address, (uint8_t)1);
 	while(bus->available() < 1);
 	return bus->read();
 }
@@ -56,7 +56,7 @@ void I2C_bus<T_I2C_bus>::read_registers(uint8_t address, uint8_t reg, uint8_t* b
 	bus->beginTransmission(address);
 	bus->write(reg);
 	bus->endTransmission();
-	bus->requestFrom(address, length);
+	bus->requestFrom((uint8_t)address, (uint8_t)length);
 	while(bus->available() < length);
 	for (uint8_t i = 0; i < length; i++) 
 	{
@@ -108,7 +108,7 @@ template <typename T_I2C_bus>
 uint8_t I2C_bus<T_I2C_bus>::receive_data(uint8_t address)
 {
 	bus->beginTransmission(address);
-	bus->requestFrom(address, (uint8_t)1);
+	bus->requestFrom((uint8_t)address, (uint8_t)1);
 	while(bus->available() < 1);
 	uint8_t value = bus->read();
 	bus->endTransmission();
@@ -119,7 +119,7 @@ template <typename T_I2C_bus>
 void I2C_bus<T_I2C_bus>::receive_data(uint8_t address, uint8_t* buffer, uint8_t length)
 {
 	bus->beginTransmission(address);
-	bus->requestFrom(address, length);
+	bus->requestFrom((uint8_t)address, (uint8_t)length);
 	while(bus->available() < length);
 	for (uint8_t i = 0; i < length; i++) 
 	{
